home/requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
filelock==3.10.7
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
packaging==23.0
parsel==1.7.0
passlib==1.7.4
pbr==5.11.1
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyDispatcher==2.0.7
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-decouple==3.8
python-multipart==0.0.6
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.7
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
starlette==0.26.1
Tempita==0.5.2
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.16.0
w3lib==2.1.1
zope.interface==6.0
